Extracted MV trail html out into MostViewedItem
